Support building with recent WAS 8.5.5 versions
To make this work, add an eclipse-plugin module that wraps the
bootstrap.jar and j2ee.jar WebSphere libraries as an OSGi bundle (for
use during the build). Those JARs are not OSGi bundles, but inside
WebSphere their classes are visible to other OSGi bundles.

Also complete the Github Actions setup to actually perform the build.

Fixes #16